agents/stock-tracker-agent/pyproject.toml
[project]
name = "stock-tracker-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk==0.2.8",
    "fastmcp==2.9.2",
    "google-adk==1.5.0",
    "google-genai>=1.17.0",
    "litellm>=1.74.15.post1",
    "python-dotenv>=1.1.0",
]